Description of the project updated
Included general known issues with the project, Description of contents
and general working of the machine. This project will no longer be updated,
apart from documentation. A New version is in course, and will be uploaded
soon. Several changes to the technologies were made, and this project no
longer supports the needs of the newer version.